requirements.txt
python-dotenv
datasets
langchain_community
langchain-openai
faiss-cpu
tiktoken
